+++ lib/databaseCommand/ChallengesDbCmd.php
@@ -1,0 +1,68 @@
+<?php
+
+namespace SimpleSAML\Module\perun\databaseCommand;
+
+/**
+ * @author Dominik Baranek <baranek@ics.muni.cz>
+ */
+class ChallengesDbCmd extends DatabaseCommand
+{
+    const CHALLENGES_TABLE_NAME = 'scriptChallenges';
+    const ID_COLUMN = 'id';
+    const CHALLENGE_COLUMN = 'challenge';
+    const SCRIPT_COLUMN = 'script';
+    const DATE_COLUMN = 'date';
+
+    public function __construct()
+    {
+        parent::__construct();
+    }
+
+    public function insertChallenge($challenge, $id, $scriptName): bool
+    {
+        $query = 'INSERT INTO ' . self::CHALLENGES_TABLE_NAME .
+            ' (' . self::ID_COLUMN . ', ' . self::CHALLENGE_COLUMN . ', ' . self::SCRIPT_COLUMN . ') VALUES' .
+            ' (:' . self::ID_COLUMN . ', :' . self::CHALLENGE_COLUMN . ', :' . self::SCRIPT_COLUMN . ')';
+
+        $params = [
+            self::ID_COLUMN => $id,
+            self::CHALLENGE_COLUMN => $challenge,
+            self::SCRIPT_COLUMN => $scriptName
+        ];
+
+        return $this->write($query, $params);
+    }
+
+    public function readChallenge($id)
+    {
+        $query = 'SELECT challenge FROM ' . self::CHALLENGES_TABLE_NAME . ' WHERE ' .
+            self::ID_COLUMN . ' = :' . self::ID_COLUMN;
+
+        $params = [
+            self::ID_COLUMN => $id
+        ];
+
+        return $this->read($query, $params)->fetchColumn();
+    }
+
+    public function deleteChallenge($id): bool
+    {
+        $query = 'DELETE FROM ' . self::CHALLENGES_TABLE_NAME . ' WHERE ' . self::ID_COLUMN . ' = :' . self::ID_COLUMN;
+
+        $params = [
+            self::ID_COLUMN => $id
+        ];
+
+        return $this->write($query, $params);
+    }
+
+    public function deleteOldChallenges(): bool
+    {
+        $query = 'DELETE FROM ' . self::CHALLENGES_TABLE_NAME . ' WHERE '
+            . self::DATE_COLUMN . ' < (NOW() - INTERVAL 5 MINUTE)';
+
+        $params = [];
+
+        return $this->write($query, $params);
+    }
+}